Move hooks into domain modules
