--- apps/template/hooks/use-scroll-contain.ts
@@ -1,59 +1,0 @@
-import { type RefObject, useEffect, useEffectEvent, useRef } from "react";
-
-function isAtScrollBoundary(el: HTMLElement, deltaY: number) {
-  const { scrollTop, scrollHeight, clientHeight } = el;
-  if (deltaY < 0) return scrollTop <= 0;
-  if (deltaY > 0) return Math.ceil(scrollTop + clientHeight) >= scrollHeight;
-  return false;
-}
-
-export function useScrollContain(
-  ref: RefObject<HTMLElement | null>,
-  enabled: boolean,
-  scrollSelector = "[data-slot=messages]",
-) {
-  const touchStartYRef = useRef(0);
-
-  const onWheel = useEffectEvent((e: WheelEvent) => {
-    const panel = ref.current;
-    if (!panel) return;
-
-    const scrollEl = panel.querySelector<HTMLElement>(scrollSelector);
-    if (scrollEl?.contains(e.target as Node) && !isAtScrollBoundary(scrollEl, e.deltaY)) {
-      return;
-    }
-    e.preventDefault();
-  });
-
-  const onTouchStart = useEffectEvent((e: TouchEvent) => {
-    touchStartYRef.current = e.touches[0]?.clientY ?? 0;
-  });
-
-  const onTouchMove = useEffectEvent((e: TouchEvent) => {
-    const panel = ref.current;
-    if (!panel) return;
-
-    const deltaY = touchStartYRef.current - (e.touches[0]?.clientY ?? 0);
-    const scrollEl = panel.querySelector<HTMLElement>(scrollSelector);
-    if (scrollEl?.contains(e.target as Node) && !isAtScrollBoundary(scrollEl, deltaY)) {
-      return;
-    }
-    e.preventDefault();
-  });
-
-  useEffect(() => {
-    if (!enabled) return;
-    const panel = ref.current;
-    if (!panel) return;
-
-    panel.addEventListener("wheel", onWheel, { passive: false });
-    panel.addEventListener("touchstart", onTouchStart, { passive: true });
-    panel.addEventListener("touchmove", onTouchMove, { passive: false });
-
-    return () => {
-      panel.removeEventListener("wheel", onWheel);
-      panel.removeEventListener("touchstart", onTouchStart);
-      panel.removeEventListener("touchmove", onTouchMove);
-    };
-  }, [enabled, ref]);
-}
