+++ apps/template/lib/search/client.ts
